# swagger1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Musicdatabase API",
    "description": "API for searching music on servers",
    "version": "1.0.0",
    "contact": {
      "name": "Yasmin",
      "email": "some-random-emaiil@some-random-domain.extension",
      "url": "http://[enter-url-here]"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development URL"
    },
    {
      "url": "https://musicdatabase1.herokuapp.com/",
      "description": "deployment site"
    }
  ],
  "tags": [
    {
      "name": "Albums",
      "description": "Manage Albums "
    },
    {
      "name": "Artist",
      "description": "Manage Artist"
    },
    { "name": "Genres", "description": "Manage Genres" },
    { "name": "Tracks", "description": "Manage Tracks" },
    { "name": "Register", "description": "Manage register" },
    { "name": "login", "description": "Manage login" }
  ],
  "externalDocs": {
    "description": "Gihub ReadMe",
    "url": "https://www.github.com/https://github.com/yasmin-pixel/musicdatabase"
  },
  "components": {
    "schemas": {
      "albums": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },

          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "example": {
          "title": "Album"
        }
      },
      "tracks": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "example": {
          "title": "Tracks"
        }
      },
      "genre": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "example": {
          "title": "Genre"
        }
      },
      "artist": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "example": {
          "title": "Artist"
        },
        "user": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "title": {
              "type": "string"
            }
          },
          "required": ["title"],
          "example": {
            "title": "User"
          }
        }
      },
      "user": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "example": {
          "title": "User"
        },
        "user": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            }
          },
          "required": ["title"],
          "example": {
            "title": "User"
          }
        }
      }
    }
  },

  "paths": {
    "/register": {
      "post": {
        "description": "register a user ",
        "tags": ["register"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "login a user ",
        "tags": ["login"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tracks/{id}": {
      "get": {
        "description": " All tracks on the database ",
        "tags": ["tracks"],
        "parameters": [
          { "name": "id", "in": "path", "description": "ID of tracks" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/albums/{id}": {
      "get": {
        "description": "albums on the database ",
        "tags": ["album"],
        "parameters": [
          { "name": "id", "in": "path", "description": "ID of album" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/artist/{id}": {
      "get": {
        "description": " All artist on database ",
        "tags": ["artist"],
        "parameters": [
          { "name": "id", "in": "path", "description": "ID of artist" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/genre": {
      "get": {
        "description": " All genre on the database ",
        "tags": ["genre"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tracks": {
      "post": {
        "description": " All tracks on the database",
        "tags": ["tracks"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "not found response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}